refactor(#579): extract RestTimerView from WorkoutTracker
Extract ~486 lines of rest timer state, logic, and template from
WorkoutTracker.vue into:
- useRestTimerController composable: all timer state (active, paused,
  seconds, presets, warnings, audio), timer control functions (start,
  stop, pause, restart), preset/warning management, and disable-with-undo
- RestTimerContent.vue: presentational component accepting a controller
  prop, rendering the timer ring, preset editor, and action buttons

WorkoutTracker retains only the modal shell, rest timer bar, and thin
coordination functions (openRestTimer, dismissTimer, skipToNextSet).
